feat: mutualize api tests
